% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{assertValidConnection}
\alias{assertValidConnection}
\title{Asserts that a connection is valid}
\usage{
assertValidConnection(con)
}
\arguments{
\item{con}{A [base::connection].}
}
\value{
Nothing or produces an informative error if the connection
is invalid.
}
\description{
Asserts that a connection is valid
}
