% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connectionId}
\alias{connectionId}
\title{Gets the internal ID of a connection}
\usage{
connectionId(con)
}
\arguments{
\item{con}{A [base::connection].}
}
\value{
A positive integer.
}
\description{
Gets the internal ID of a connection
}
