% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{connectionInfo}
\alias{connectionInfo}
\title{Gets a string summary of a connection}
\usage{
connectionInfo(con)
}
\arguments{
\item{con}{A [base::connection].}
}
\value{
A character string.
}
\description{
Gets a string summary of a connection
}
