% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{isValidConnection}
\alias{isValidConnection}
\title{Checks whether a connection is valid or not}
\usage{
isValidConnection(con)
}
\arguments{
\item{con}{A [base::connection].}
}
\value{
Returns TRUE if the connection is valid, otherwise FALSE.
If FALSE, the attribute `reason` explains why it is invalid.
}
\description{
Checks whether a connection is valid or not
}
